# docs/swagger.yaml
definitions:
  jsonformatter.LeaseResponse:
    properties:
      created_at:
        example: 2025-11-03T13:45:00+03
        type: string
      node_id:
        example: some-node-id
        type: string
      updated_at:
        example: 2025-11-03T13:45:00+03
        type: string
      workload_id:
        example: some-workload-id
        type: string
    type: object
  jsonformatter.WorkloadResponse:
    properties:
      id:
        example: some-uuid
        type: string
      status:
        example: new
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1//workloads/{workload_id}/lease:
    post:
      consumes:
      - application/json
      description: Create a lease for a specific workload on a given node
      parameters:
      - description: Workload ID
        in: path
        minLength: 1
        name: workload_id
        required: true
        type: string
      - description: Node ID
        in: query
        name: node_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/jsonformatter.LeaseResponse'
        "400":
          description: Invalid request parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a lease
      tags:
      - leases
    put:
      consumes:
      - application/json
      description: Extend a lease for a specific workload on a given node
      parameters:
      - description: Workload ID
        in: path
        minLength: 1
        name: workload_id
        required: true
        type: string
      - description: Node ID
        in: query
        name: node_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid request parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Extend a lease
      tags:
      - leases
  /api/v1/workloads:
    get:
      consumes:
      - application/json
      description: Retrieve a list of workloads filtered by status, CPU, and RAM
      parameters:
      - description: Filter by workload status
        example: new
        in: query
        name: status
        type: string
      - description: Filter by CPU (millicores)
        example: 100
        in: query
        name: cpu
        type: integer
      - description: Filter by RAM (MB)
        example: 256
        in: query
        name: ram
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/jsonformatter.WorkloadResponse'
              type: array
            type: array
        "400":
          description: Invalid request parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List workloads
      tags:
      - workloads
swagger: "2.0"